moved rclone setting sto director-v2
