+++ services/director-v2/tests/integration/02/utils.py
@@ -2,9 +2,9 @@
 
 import asyncio
 import json
+import logging
 import os
 from typing import Any, Dict, Optional, Set
-import logging
 
 import aiodocker
 import httpx
