+++ services/director-v2/tests/unit/test_core_settings.py
@@ -2,16 +2,55 @@
 # pylint:disable=unused-argument
 # pylint:disable=redefined-outer-name
 
-from typing import Any, Dict
+from typing import Any, Dict, Set
 
 import pytest
+from _pytest.monkeypatch import MonkeyPatch
 from models_library.basic_types import LogLevel
 from simcore_service_director_v2.core.settings import (
     AppSettings,
     BootModeEnum,
     DynamicSidecarProxySettings,
     DynamicSidecarSettings,
+    RCloneSettings,
+    S3Provider,
 )
+
+
+def _get_backend_type_options() -> Set[str]:
+    return {x for x in dir(S3Provider) if not x.startswith("_")}
+
+
+def test_supported_backends_did_not_change() -> None:
+    _EXPECTED = {"AWS", "CEPH", "MINIO"}
+    assert _EXPECTED == _get_backend_type_options(), (
+        "Backend configuration change, please code support for "
+        "it in volumes_resolver -> _get_s3_volume_driver_config. "
+        "When done, adjust above list."
+    )
+
+
+@pytest.mark.parametrize(
+    "endpoint, is_secure",
+    [
+        ("localhost", False),
+        ("s3_aws", True),
+        ("https://ceph.home", True),
+        ("http://local.dev", False),
+    ],
+)
+def test_expected_endpoint(
+    endpoint: str, is_secure: bool, monkeypatch: MonkeyPatch
+) -> None:
+    monkeypatch.setenv("S3_PROVIDER", "MINIO")
+    monkeypatch.setenv("S3_ENDPOINT", endpoint)
+    monkeypatch.setenv("S3_SECURE", "true" if is_secure else "false")
+
+    r_clone_settings = RCloneSettings()
+
+    scheme = "https" if is_secure else "http"
+    assert r_clone_settings.endpoint.startswith(f"{scheme}://")
+    assert r_clone_settings.endpoint.endswith(endpoint)
 
 
 def test_settings_with_project_env_devel(project_env_devel_environment: Dict[str, Any]):
